requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
httptools==0.6.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2==2.9.8
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3

googletrans==3.1.0a0

fastapi-cache2